gptsovits_service/requirements.txt
fastapi
uvicorn
numpy
soundfile

python-dotenv
psutil
